+++ common-utils/src/main/java/net/anweisen/utilities/commons/common/SeededRandomWrapper.java
@@ -1,4 +1,4 @@
-package net.anweisen.utilities.commons.misc;
+package net.anweisen.utilities.commons.common;
 
 import java.util.Random;
